# Vision
Jonathan Aquino [suggests][1] that version control commit messages can
be a form of Literate Programming.

[1]: http://jona.ca/blog/git-commit-messages-are-the-new-literate-programming

The original Literate Programming concept by Donald Knuth suggested that
the code should be embedded in comments describing everything that was
in the head of the author while the code was being written. This never
caught on, perhaps because this leads to code with too long comments.

On the other hand, Jonathan notes that some people write really detailed
Git commit messages.

Would it be feasible to present a Git repository as a collection of
chapters and sections, with each coherent change accompanied with a
descriptive commit message telling the story behind the change?

Using the GitHub API, we can extract commit messages from a repository
and render them as Markdown documents, along with diffs.

I imagine that composing such a repository requires a vastly different
discipline and/or workflow than many of us are used to.

* For example, you don’t just `git commit -am 'Fix bug'`; you actually
start writing your commit message _before_ you start fixing the bug,
while investigating the bug cause, the workarounds, and the solution.
When implementing a new feature, you might start with an empty commit,
explaining the need for the feature and/or its design (like I’m doing
here).
* I also think this requires a good deal of history rewriting. For
example, just now I had to fix the syntax of the hyperlink above, so
that it would be parsed by the Markdown library I chose. This caused a
rebase of the whole repository.