Display repository description
Let’s try interacting with the GitHub API. As usual, we create an
`XMLHttpRequest`, hook up an event handler, and send the request.

For now, I hardcoded the repository owner and name. In future, they can
be taken from the URL query string.

Also, for simplicity, I treat the description as plain text; I will get
to Markdown formatting later.