Factor out AJAX mechanics
Before we display the list of commits, notice that a single API request
requires quite a few lines of code. Let’s factor out all the repeating
steps in a function.

The two things that change from one API call to another are the URL and
the handler function. However, the handler will always need to parse a
JSON response.